+++ src/messages/index.ts
@@ -1,10 +1,37 @@
 import { ICourierClientConfiguration } from "../types";
-import { ICourierClientMessages, ICourierMessageGetResponse } from "./types";
+import {
+  ICourierClientMessages,
+  ICourierMessage,
+  ICourierMessageGetAllResponse,
+  ICourierMessageHistory
+} from "./types";
 
 export const getMessage = (options: ICourierClientConfiguration) => {
-  return async (messageId: string): Promise<ICourierMessageGetResponse> => {
-    const res = await options.httpClient.get<ICourierMessageGetResponse>(
+  return async (messageId: string): Promise<ICourierMessage> => {
+    const res = await options.httpClient.get<ICourierMessage>(
       `/messages/${messageId}`
+    );
+    return res.data;
+  };
+};
+
+const getMessages = (options: ICourierClientConfiguration) => {
+  return async (params?: {
+    cursor: string;
+    recipient: string;
+  }): Promise<ICourierMessageGetAllResponse> => {
+    const res = await options.httpClient.get<ICourierMessageGetAllResponse>(
+      `/messages`,
+      params
+    );
+    return res.data;
+  };
+};
+
+const getMessageHistory = (options: ICourierClientConfiguration) => {
+  return async (messageId: string): Promise<ICourierMessageHistory> => {
+    const res = await options.httpClient.get<ICourierMessageHistory>(
+      `/messages/${messageId}/history`
     );
     return res.data;
   };
@@ -13,5 +40,9 @@ export const getMessage = (options: ICourierClientConfiguration) => {
 export const messages = (
   options: ICourierClientConfiguration
 ): ICourierClientMessages => {
-  return { getMessage: getMessage(options) };
+  return {
+    getMessage: getMessage(options),
+    getMessageHistory: getMessageHistory(options),
+    getMessages: getMessages(options)
+  };
 };