+++ src/messages/types.ts
@@ -1,13 +1,24 @@
-export interface ICourierMessageGetResponse {
+import { ICourierPaging } from "../types";
+
+export interface ICourierMessage {
+  clicked?: number;
+  delivered?: number;
+  error?: string;
   enqueued?: number;
   event?: string;
   id: string;
   notification?: string;
+  opened?: number;
   providers?: Array<{
     channel: {
+      key: string;
       name: string;
       template: string;
     };
+    clicked?: number;
+    delivered?: number;
+    error?: string;
+    opened?: number;
     provider: string;
     reference: {
       "x-message-id": string;
@@ -15,11 +26,26 @@ export interface ICourierMessageGetResponse {
     sent: number;
     status: string;
   }>;
+  reason?: string;
   recipient: string;
   sent?: number;
   status: string;
 }
 
+export interface ICourierMessageHistory {
+  results: object;
+}
+
+export interface ICourierMessageGetAllResponse {
+  paging: ICourierPaging;
+  results: ICourierMessage[];
+}
+
 export interface ICourierClientMessages {
-  getMessage: (messageId: string) => Promise<ICourierMessageGetResponse>;
+  getMessage: (messageId: string) => Promise<ICourierMessage>;
+  getMessageHistory: (messageId: string) => Promise<ICourierMessageHistory>;
+  getMessages: (params?: {
+    cursor: string;
+    recipient: string;
+  }) => Promise<ICourierMessageGetAllResponse>;
 }